File reorganization, tooltip improvements
