+++ modules/mob_types/group_types/caravans.py
@@ -2,12 +2,12 @@
 
 import random
 import math
-from .groups import group
-from .. import utility
-from .. import actor_utility
-from .. import dice_utility
-from .. import notification_tools
-from .. import market_tools
+from ..groups import group
+from ... import utility
+from ... import actor_utility
+from ... import dice_utility
+from ... import notification_tools
+from ... import market_tools
 
 class caravan(group):
     '''
