+++ modules/mob_types/group_types/construction_gangs.py
@@ -1,9 +1,9 @@
 #Contains functionality for construction gangs
 
-from .groups import group
-from .. import actor_utility
-from .. import dice_utility
-from .. import notification_tools
+from ..groups import group
+from ... import actor_utility
+from ... import dice_utility
+from ... import notification_tools
 
 class construction_gang(group):
     '''
