+++ modules/mob_types/group_types/expeditions.py
@@ -2,12 +2,12 @@
 
 import time
 import random
-from .groups import group
-from ..tiles import tile
-from .. import actor_utility
-from .. import text_tools
-from .. import dice_utility
-from .. import notification_tools
+from ..groups import group
+from ...tiles import tile
+from ... import actor_utility
+from ... import text_tools
+from ... import dice_utility
+from ... import notification_tools
 
 class expedition(group):
     '''
