+++ modules/mob_types/group_types/porters.py
@@ -1,7 +1,7 @@
 #Contains functionality for porters
 
-from .groups import group
-from .. import actor_utility
+from ..groups import group
+from ... import actor_utility
 
 class porters(group):
     '''
